Order improved

+++ database/migrations/2026_07_31_090428_create_orders_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
             $table->foreignId('cart_id')->constrained()->onDelete('cascade');
             $table->enum('status', ['pending', 'paid', 'cancelled'])->default('pending');
-            $table->float('total_price');
+            $table->unsignedBigInteger('total_price');
             $table->timestamps();
         });
     }
